make new file for keyboard shortcuts
