+++ Sources/CodeEditSourceEditor/Controller/TextViewController+LoadView.swift
@@ -109,5 +109,19 @@ extension TextViewController {
                 }
             }
             .store(in: &cancellables)
+
+
+        NSEvent.addLocalMonitorForEvents(matching: .keyDown) { event in
+            guard self.view.window?.firstResponder == self.textView else { return event }
+            let charactersIgnoringModifiers = event.charactersIgnoringModifiers
+            let commandKey = NSEvent.ModifierFlags.command.rawValue
+            let modifierFlags = event.modifierFlags.intersection(.deviceIndependentFlagsMask).rawValue
+            if modifierFlags == commandKey && event.charactersIgnoringModifiers == "/" {
+                self.commandSlashCalled()
+                return nil
+            } else {
+                return event
+            }
+        }
     }
 }
